pyproject.toml
[project]
name = "prompt-to-database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.51",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.12",
    "mysql-connector-python>=9.2.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
